requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.26.0
pyyaml>=6.0.1
aiofiles>=23.2.1
